Fixed shape of momentum variables P.
When calculating the ZMP, chain_model.P returned a size (3,3) matrix
when it should return an array of size 3. The ZMP unit test did not
not catch this because P happend to be all zeros, which made its
effect on subsequent calculations irrelevant.

update_chain_dynamics.m:
- chain_model.joint(idx_to).v was transposed to
match the shape of the cross product that it will be summed with.
- shape of chain_model.joint(idx_to).w has been made uniform to remove
unnecessary transposes and since the cross product automatically
identifies the dimension with size 3